+++ packages/babel-plugin/src/replaceK.ts
@@ -1,4 +1,7 @@
 import { NodePath, Node, types, template } from "@babel/core";
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import printAST from "ast-pretty-print";
 
 /**
  * Processes the CallExpression nodes in the AST and replaces the 'k' function from '@kuma-ui/core'
@@ -15,30 +18,25 @@ export const replaceK = (
   importedStyleFunctions: Record<string, string>
 ) => {
   node.traverse({
-    CallExpression(path) {
-      const { node } = path;
+    JSXElement(path) {
+      const { openingElement, closingElement } = path.node;
 
       if (
-        node.callee.type === "MemberExpression" &&
-        t.isIdentifier(node.callee.object) &&
-        node.callee.object.name === "React" &&
-        node.callee.property.type === "Identifier" &&
-        (node.callee.property.name === "createElement" ||
-          node.callee.property.name === "cloneElement")
+        t.isJSXMemberExpression(openingElement.name) &&
+        t.isJSXIdentifier(openingElement.name.object, {
+          name: importedStyleFunctions["k"],
+        })
       ) {
-        const firstArg = path.get("arguments")[0];
-
         if (
-          t.isMemberExpression(firstArg.node) &&
-          t.isIdentifier(firstArg.node.object) &&
-          firstArg.node.object.name === importedStyleFunctions["k"]
+          closingElement &&
+          t.isJSXMemberExpression(closingElement.name) &&
+          t.isJSXIdentifier(closingElement.name.object, {
+            name: importedStyleFunctions["k"],
+          })
         ) {
-          let htmlTag = t.isIdentifier(firstArg.node.property)
-            ? firstArg.node.property.name
-            : undefined;
-          htmlTag ||= "div";
-          firstArg.replaceWith(t.stringLiteral(htmlTag));
+          closingElement.name = openingElement.name.property;
         }
+        openingElement.name = openingElement.name.property;
       }
     },
   });