+++ packages/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kuma-ui/core
 
+## 0.4.2
+
+### Patch Changes
+
+- 8af4f8d: chore: Automate synchronization of README.md across packages.
+- Updated dependencies [8af4f8d]
+  - @kuma-ui/system@0.6.2
+
+## 0.4.1
+
+### Patch Changes
+
+- 558823e: fix classname hash
+  - @kuma-ui/system@0.6.1
+
 ## 0.4.0
 
 ### Minor Changes