+++ packages/system/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kuma-ui/system
 
+## 0.6.2
+
+### Patch Changes
+
+- 8af4f8d: chore: Automate synchronization of README.md across packages.
+- Updated dependencies [8af4f8d]
+  - @kuma-ui/sheet@0.5.1
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [558823e]
+  - @kuma-ui/sheet@0.5.0
+
 ## 0.6.0
 
 ### Minor Changes